Compiler warning
git-svn-id: file:///home/michael/svn/yuni/trunk@1622 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/tools/yuni-docmake/job-writer.cpp
@@ -111,7 +111,7 @@ namespace // anonymous
 		if (pToDelete.empty())
 			return;
 		// start from the last item to delete
-		unsigned int i = pToDelete.size();
+		unsigned int i = (unsigned int) pToDelete.size();
 		do
 		{
 			--i;
